requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.13.0=pypi_0
astunparse=1.6.3=pypi_0
bcolz=1.2.1=py38h962f231_0
blas=1.0=mkl
blosc=1.21.0=h8c45485_0
bokeh=2.3.3=py38h06a4308_0
bottleneck=1.3.2=py38heb32a55_1
brotli=1.0.9=he6710b0_2
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.7.5=h06a4308_1
cachetools=4.2.2=pypi_0
certifi=2021.5.30=py38h06a4308_0
charset-normalizer=2.0.4=pypi_0
clang=5.0=pypi_0
click=8.0.1=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
cudatoolkit=11.0.3=h15472ef_8
cycler=0.10.0=py38_0
cytoolz=0.11.0=py38h7b6447c_0
dask=2021.8.0=pyhd3eb1b0_0
dask-core=2021.8.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
distributed=2021.8.0=py38h06a4308_0
expat=2.4.1=h2531618_2
flatbuffers=1.12=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.7.0=pyhd3eb1b0_0
future=0.18.2=py38_1
gast=0.4.0=pypi_0
glib=2.69.0=h5202010_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.5=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.39.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.1.0=pypi_0
hdf5=1.10.4=hb1b8bf9_0
heapdict=1.0.1=py_0
icu=58.2=he6710b0_3
idna=3.2=pypi_0
importlib-metadata=3.10.0=py38h06a4308_0
intel-openmp=2021.3.0=h06a4308_3350
jinja2=3.0.1=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
keras=2.6.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py38h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
locket=0.2.1=py38h06a4308_1
lz4-c=1.9.3=h295c915_1
lzo=2.10=h7b6447c_2
markdown=3.3.4=pypi_0
markupsafe=2.0.1=py38h27cfd23_0
matplotlib=3.4.2=py38h06a4308_0
matplotlib-base=3.4.2=py38hab158f2_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
mock=4.0.3=pyhd3eb1b0_0
msgpack-python=1.0.2=py38hff7bd54_1
munkres=1.1.4=py_0
ncurses=6.2=he6710b0_1
ninja=1.10.2=hff7bd54_1
numexpr=2.7.3=py38h22e1b3c_1
numpy=1.19.5=pypi_0
oauthlib=3.1.1=pypi_0
olefile=0.46=py_0
opencv-python=4.5.3.56=pypi_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h27cfd23_0
opt-einsum=3.3.0=pypi_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.3.1=py38h8c16a72_0
partd=1.2.0=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pillow=8.3.1=py38h2c7a002_0
pip=21.0.1=py38h06a4308_0
protobuf=3.17.3=pypi_0
psutil=5.8.0=py38h27cfd23_1
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pytables=3.6.1=py38h9fd0a39_0
python=3.8.11=h12debd9_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.7.0=py3.8_cuda11.0.221_cudnn8.0.3_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=py38h27cfd23_1
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-learn=0.24.2=py38ha9443f7_0
scipy=1.6.2=py38had2a1c9_1
seaborn=0.11.2=pypi_0
setuptools=52.0.0=py38h06a4308_0
sip=4.19.13=py38he6710b0_0
six=1.15.0=pypi_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tblib=1.7.0=py_0
tensorboard=2.6.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorboardx=2.4=pypi_0
tensorflow-estimator=2.6.0=pypi_0
tensorflow-gpu=2.6.0=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=2.2.0=pyhbf3da8f_0
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torch-tb-profiler=0.2.1=pypi_0
torchaudio=0.7.0=py38
torchvision=0.8.0=py38_cu110
tornado=6.1=py38h27cfd23_0
tqdm=4.62.0=pyhd3eb1b0_1
typing-extensions=3.7.4.3=pypi_0
urllib3=1.26.6=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.37.0=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zict=2.0.0=pyhd3eb1b0_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
